update: request validation exception handling
